Move AppWishListButton component to matters dir
